requirements.txt
typer
rich
python-dotenv
groq
google-genai